pyproject.toml
[project]
name = "toki"
version = "0.1.0"
description = "Toki - A Photo Library Management Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
    "hachoir>=3.0.0",
]

[project.scripts]
toki = "toki.main:cli"
